# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sistema de Reservas de Restaurantes",
        "description": "API para la gestión de clientes, mesas y reservas.",
        "contact": {
            "name": "Vladimir Faundez",
            "email": "v.faundezh@gmail.com"
        },
        "license": {
            "name": "Creative Commons Attribution-NonCommercial-NoDerivatives 4.0 International",
            "url": "https://creativecommons.org/licenses/by-nc-nd/4.0/legalcode"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor principal de la API"
        }
    ],
    "paths": {
...
        "/api/v1/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get all reservation",
                "description": "Retrieve a list of all reservations.",
                "operationId": "getAllReservations",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReservationSuccess"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Create a new reservation",
                "description": "Store a new reservation in the system.",
                "operationId": "storeReservation",
                "requestBody": {
                    "$ref": "#/components/requestBodies/StoreReservationRequest"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/ReservationSuccess"
                    },
                    "400": {
                        "$ref": "#/components/responses/ReservationInvalidRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{reservation}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Get a reservation by ID",
                "description": "Retrieve a specific reservation by its ID.",
                "operationId": "getReservationById",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID of the reservation to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReservationSuccessId"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Update a reservation",
                "description": "Update an existing reservation in the system.",
                "operationId": "updateReservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID of the reservation to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateReservationRequest"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/ReservationSuccessId"
                    },
                    "400": {
                        "$ref": "#/components/responses/ReservationInvalidRequest"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Delete a reservation",
                "description": "Delete a specific reservation by its ID.",
                "operationId": "deleteReservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID of the reservation to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation deleted successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{reservation}/completed": {
            "patch": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Complete a reservation",
                "description": "Mark a specific reservation as completed.",
                "operationId": "completeReservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID of the reservation to complete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation completed successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/reservations/{reservation}/cancelled": {
            "patch": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Cancel a reservation",
                "description": "Mark a specific reservation as cancelled.",
                "operationId": "cancelReservation",
                "parameters": [
                    {
                        "name": "reservation",
                        "in": "path",
                        "description": "ID of the reservation to cancel",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reservation cancelled successfully"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthenticatedError"
                    },
                    "403": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundError"
                    },
                    "500": {
                        "$ref": "#/components/responses/ServerError"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Reservation": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "reservation_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-01-01"
                    },
                    "reservation_time": {
                        "type": "string",
                        "format": "time",
                        "example": "12:00"
                    },
                    "number_of_people": {
                        "type": "integer",
                        "example": 4
                    },
                    "status": {
                        "type": "string",
                        "example": "confirmed"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Family dinner"
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 3
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "UnauthenticatedResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    },
                    "data": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "UnauthorizedResponse": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "data": {
                        "type": "string",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "NotFoundError": {
                "description": "Reservation not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFoundResponse"
                        }
                    }
                }
            },
            "UnauthenticatedError": {
                "description": "Unauthenticated user",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthenticatedResponse"
                        }
                    }
                }
            },
            "UnauthorizedError": {
                "description": "Unauthorized action",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UnauthorizedResponse"
                        }
                    }
                }
            },
            "ServerError": {
                "description": "Internal server error",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/ServerErrorResponse"
                        }
                    }
                }
            },
            "ReservationSuccess": {
                "description": "Successful reservation operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Reservation"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ReservationSuccessId": {
                "description": "Successful reservation operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": ""
                                },
                                "data": {
                                    "$ref": "#/components/schemas/Reservation"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ReservationInvalidRequest": {
                "description": "Invalid request data",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Invalid request"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "field": {
                                                "type": "string",
                                                "example": "name"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "The name field is required."
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
...
        "requestBodies": {
...
            "StoreReservationRequest": {
                "description": "Data required to create a reservation",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "reservation_date",
                                "reservation_time",
                                "number_of_people",
                                "status",
                                "notes",
                                "customer_id",
                                "user_id",
                                "tables"
                            ],
                            "properties": {
                                "reservation_date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2024-01-01"
                                },
                                "reservation_time": {
                                    "type": "string",
                                    "format": "time",
                                    "example": "12:00"
                                },
                                "number_of_people": {
                                    "type": "integer",
                                    "example": 4
                                },
                                "status": {
                                    "type": "string",
                                    "example": "confirmed"
                                },
                                "notes": {
                                    "type": "string",
                                    "example": "Family dinner"
                                },
                                "customer_id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "user_id": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "tables": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    },
                                    "example": [
                                        1,
                                        2
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "UpdateReservationRequest": {
                "description": "Data required to update a reservation",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "required": [
                                "reservation_date",
                                "reservation_time",
                                "number_of_people",
                                "status",
                                "notes",
                                "customer_id",
                                "user_id",
                                "tables"
                            ],
                            "properties": {
                                "reservation_date": {
                                    "type": "string",
                                    "format": "date",
                                    "example": "2024-01-01"
                                },
                                "reservation_time": {
                                    "type": "string",
                                    "format": "time",
                                    "example": "12:00"
                                },
                                "number_of_people": {
                                    "type": "integer",
                                    "example": 4
                                },
                                "status": {
                                    "type": "string",
                                    "example": "confirmed"
                                },
                                "notes": {
                                    "type": "string",
                                    "example": "Family dinner"
                                },
                                "customer_id": {
                                    "type": "integer",
                                    "example": 3
                                },
                                "user_id": {
                                    "type": "integer",
                                    "example": 2
                                },
                                "tables": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer"
                                    },
                                    "example": [
                                        1,
                                        2
                                    ]
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "description": "JWT Authorization header usando el esquema Bearer.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}